+++ Changelog.md
@@ -4,3 +4,4 @@
 
 - rendering all `UserControl`s implementing the `SettingsEditorItem` interface
 - loading, validating and saving all rendered settings
+- running a reload of the whole `SettingsEditor`